% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_compare.R
\name{compare}
\alias{compare}
\title{compare two clusters}
\usage{
compare(clustering, alt.clustering)
}
\arguments{
\item{clustering}{an integer vector of length of the number of cases, which indicates a clustering}

\item{alt.clustering}{an integer vector such as for clustering, indicating an alternative clustering}
}
\value{
jaccard

rand

ajusted.rand

vi

nmi
}
\description{
compare two clusters
}
\examples{
data(iris)
X <- scale(iris[,1:4])
c1 <- kmeans(X, 3)
c2 <- kmeans(X, 4)
compare(c1$cluster, c2$cluster)
}

