% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareClustering.R
\name{compareClustering}
\alias{compareClustering}
\title{Parallel Validation Of Clusteranalyses}
\usage{
compareClustering(dataMatrix, maxClusters, distanceMeasures = c("euclidean",
  "manhattan"), clusteringMethods = c("ward.D2", "single", "complete",
  "average", "mcquitty", "diana", "kmeans"), sfParallel = TRUE, sfCpus = 2,
  ...)
}
\arguments{
\item{dataMatrix}{a data matrix accepted by \code{stats::\link[stats]{dist}()}}

\item{maxClusters}{the maximum number of clusters to evaluate}

\item{distanceMeasures}{a character vector of the distance measures to use
(currently, only "\code{euclidean}", "\code{maximum}",
"\code{manhattan}", "\code{canberra}", "\code{binary}" or "\code{minkowski}" are allowed)}

\item{clusteringMethods}{a character vector of cluster methods to use (currently, the following are allowed:
\itemize{
    \item{\code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"}
    , \code{"average"}, \code{"mcquitty"}, \code{"median"}, \code{"centroid"}
    (use \code{stats::\link[stats]{hclust}})}
    \item{\code{"diana"} (uses \code{cluster::\link[cluster]{diana}})}
    \item{\code{"kmeans"} (uses \code{cluster::\link[cluster]{pam}})}
})}

\item{sfParallel}{logical Should \code{\link[snowfall]{snowfall}} be used in parallel mode?}

\item{sfCpus}{number of cpu to use}

\item{...}{passed to \code{snowfall::\link[snowfall]{sfInit}()}}
}
\value{
a \code{tibble::\link[tibble]{tibble}} with one row per distance measure,
    method and number of clusters
    from 2 to \code{k} and the columns:
    \describe{
        \item{distance}{= \code{dm}}
        \item{method}{= \code{method}}
        \item{nCluster}{= \code{k}}
        \item{totalAvgSilWidth}{overall average silhoutte width (\code{cluster::\link[cluster]{summary.silhouette}$avg.width})}
        \item{minClustAvgSilWidth}{minimal average cluster silhoutte width (\code{cluster::\link[cluster]{summary.silhouette}$clus.avg.widths})}
        \item{minSilWidth}{minimal silhoutte width (\code{cluster::\link[cluster]{summary.silhouette}$si.summary$`Min.`})}
        \item{pPosSilWidths}{percentage of positive silhoutte widths}
        \item{minClustJacMean}{minimal cluster bootstrap mean of Jaccard's index (\code{fpc::\link[fpc]{clusterboot}$bootmean})}
        \item{pClustJacOver06}{percentage of cluster bootstrap means of Jaccard's index above 0.6}
        \item{separationIndex}{\code{fpc::\link[fpc]{cluster.stats}$sindex}}
        \item{avgDistWithin}{\code{fpc::\link[fpc]{cluster.stats}$average.within}}
        \item{withinVsBetween}{\code{fpc::\link[fpc]{cluster.stats}$wb.ratio}}
    }
}
\description{
Parallel Validation Of Clusteranalyses
}
