% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dianaCBI.R
\name{dianaCBI}
\alias{dianaCBI}
\title{Clusterboot Interface To Diana}
\usage{
dianaCBI(x, k)
}
\arguments{
\item{x}{a distance matrix}

\item{k}{the number of clusters}
}
\value{
A \code{list} containing the following components:
\describe{
    \item{result}{an object of class "\code{diana}",
    the output of the call to \code{cluster::\link[cluster]{diana}()}}
    \item{nc}{the number of clusters (\code{k})}
    \item{clusterlist}{a list consisting of logical vectors of length of
    the number of data points for each cluster indicating whether a point is a member of this cluster}
    \item{partition}{an integer vector of length of
    the number of data points partitioning the data}
    \item{clustermethod}{"\code{cluster::diana}"}
}
}
\description{
Provides an interface to \code{cluster::\link[cluster]{diana}()} that allows the execution
from \code{fpc::\link[fpc]{clusterboot}()}.
}
\seealso{
\code{fpc::\link[fpc]{clusterboot}()}
}
