% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateClusters.R
\name{evaluateClusters}
\alias{evaluateClusters}
\title{Validating A Clusteranalysis}
\usage{
evaluateClusters(xList)
}
\arguments{
\item{xList}{a list with the following elements:
\describe{
    \item{dm}{a string describing a distance measure (currently, only "\code{euclidean}", "\code{maximum}",
    "\code{manhattan}", "\code{canberra}", "\code{binary}" or "\code{minkowski}" are allowed)}
    \item{method}{a string naming a clustering method (currently, the following are allowed:
    \itemize{
        \item{\code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"}
        , \code{"average"}, \code{"mcquitty"}, \code{"median"}, \code{"centroid"}
        (use \code{stats::\link[stats]{hclust}})}
        \item{\code{"diana"} (uses \code{cluster::\link[cluster]{diana}})}
        \item{\code{"kmeans"} (uses \code{cluster::\link[cluster]{pam}})}
    }
    )}
    \item{k}{the maximum number of clusters}
    \item{distData}{a distance matrix}
}}
}
\value{
a \code{tibble::\link[tibble]{tibble}} with one row per number of clusters
    from 2 to \code{k} and the columns:
    \describe{
        \item{distance}{= \code{dm}}
        \item{method}{= \code{method}}
        \item{nCluster}{= \code{k}}
        \item{totalAvgSilWidth}{overall average silhoutte width (\code{cluster::\link[cluster]{summary.silhouette}$avg.width})}
        \item{minClustAvgSilWidth}{minimal average cluster silhoutte width (\code{cluster::\link[cluster]{summary.silhouette}$clus.avg.widths})}
        \item{minSilWidth}{minimal silhoutte width (\code{cluster::\link[cluster]{summary.silhouette}$si.summary$`Min.`})}
        \item{pPosSilWidths}{percentage of positive silhoutte widths}
        \item{minClustJacMean}{minimal cluster bootstrap mean of Jaccard's index (\code{fpc::\link[fpc]{clusterboot}$bootmean})}
        \item{pClustJacOver06}{percentage of cluster bootstrap means of Jaccard's index above 0.6}
        \item{separationIndex}{\code{fpc::\link[fpc]{cluster.stats}$sindex}}
        \item{avgDistWithin}{\code{fpc::\link[fpc]{cluster.stats}$average.within}}
        \item{withinVsBetween}{\code{fpc::\link[fpc]{cluster.stats}$wb.ratio}}
    }
}
\description{
Called by \code{\link{compareClustering}()}.
}
