% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_explore.R
\name{explore}
\alias{explore}
\title{explore data for clustering}
\usage{
explore(formula, data, classifier = = c("randomForest", "ranger"), control = NULL)
}
\description{
explore data to detect whether it can be splited into proper clusters and find which variables can be used to cluster
}
\examples{
data(iris)
iris1 <- iris[iris$Species == "versicolor",1:4]
e1 <- explore(~., iris1, "randomForest")
summary(e1) # there are no clusters in the data
plot(e1)

e2 <- explore(~., iris[,1:4], "ranger")
summary(e2)
plot(e2)
}

