% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_seeds.R
\name{seeding}
\alias{seeding}
\title{finding initial cluster centers}
\usage{
seeding(X, num_seeds, m, threads)
}
\arguments{
\item{X}{a numeric matrix or data frame}

\item{num_seeds}{number of seeds}

\item{m}{an integer, length of Markove chains, which is used to sample centers, 20 by default}

\item{threads}{an integer, number of threads to speed up computing}
}
\description{
A fast seeding algorithm proposing to approximate k-means++
using Markov chain Monte Carlo (MCMC).
}
\examples{
data(iris)
iris <- iris[sample(1:nrow(iris), 5000, replace = T),]
X <- as.matrix(iris[,1:4])
seeds <- seeding(X, 3, 20, 2)
clus1 <- kmeans(X, X[seeds,]); table(clus1$cluster, iris[,5])
clus2 <- kmeans(X, 3); table(clus2$cluster, iris[,5])
}

